refactor(airc-lib+cli): extract GhClient trait to airc-lib — ShellGhClient stays in cli (card a094aa81)
The trait, error enum, arg/result types, and pure parse helpers
move to `airc_lib::gh_client`; the `tokio::process::Command::new("gh")`
spawn stays in `airc_cli::gh_client` as `ShellGhClient`. The
merger threads `&dyn GhClient` end-to-end so the impl is now
substitutable without touching the gate logic.

Why now: card a094aa81 (substrate-vision consumer embedding).
Stacked on top of #1041 (card dec35ec7) which introduced the
concrete struct. Continuum/OpenClaw/Hermes/codex can pull
airc-lib and get the gh boundary without linking the entire
CLI binary; the shell-based impl becomes the airc CLI's
choice of backend, not an obligation imposed on consumers.

API shape: args structs now own their strings (PathBuf/String
instead of &Path/&str). Tradeoff: a few extra .to_string()
calls at construction; in return, #[async_trait] desugars
cleanly and the trait can be stored as Arc<dyn GhClient> for
mocking. The call frequency is "merger tick" and "pr create" —
allocation cost is noise vs. the gh spawn itself.

Validation:
- workspace builds clean
- airc-lib::gh_client::tests = 5/5 (parse_pr_view + parse_pr_url)
- airc-cli::gh_client::tests = 6/6 (map_spawn_error + classify_*)
- airc-cli bin tests = 167/167
- cargo fmt --check clean
- cargo clippy --workspace --all-targets -D warnings clean

Stacked-PR note: branch is based on #1041 (dec35ec7). When that
merges to rust-rewrite, GitHub auto-rebases this PR onto the
single new commit.
