+++ crates/airc-cli/Cargo.toml
@@ -44,6 +44,7 @@ rand_core = { version = "0.6", features = ["getrandom"] }
 fs2 = "0.4"
 chrono = { version = "0.4", default-features = false, features = ["clock", "std"] }
 thiserror = "1"
+async-trait = "0.1"
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
