+++ crates/airc-lib/src/lib.rs
@@ -45,6 +45,7 @@ pub mod diagnostic_event_sink;
 pub mod error;
 pub mod external_identity;
 pub mod gh_account_registry;
+pub mod gh_client;
 pub mod join_context;
 mod lan;
 pub mod lane_coordination;
@@ -102,6 +103,10 @@ pub use external_identity::{
     HEADER_BRIDGE_HANDLE, HEADER_BRIDGE_SOURCE,
 };
 pub use gh_account_registry::{gh_auth_ready, GhAccountRegistryStore};
+pub use gh_client::{
+    parse_pr_url, parse_pr_view, GhCheck, GhClient, GhError, MergeReceipt, PrCreateArgs, PrCreated,
+    PrEditBaseArgs, PrMergeArgs, PrView, PrViewArgs,
+};
 pub use join_context::{JoinContext, GENERAL_CHANNEL};
 pub use lane_coordination::{
     LaneAction, LaneCoordinationEvent, LaneStatus, HEADER_COORD_KIND, HEADER_COORD_LANE_ID,
